Fixed occasional crash caused by threading/timing issue with ReminderAlarm#onReceive (closes #591).
ReminderAlarm#onReceive can be triggered after UserEventManager sets its UserEvent to null. What probably happens: AlarmManager is in the process of firing the next ReminderAlarm in an OS thread when we call "cancelAlarm" in our thread. Since it has already started its firing process, ReminderAlarm#onReceive gets called and we dereference a null UserEvent pointer.

I've also moved the call to cancelAlarm() in UserEventManager because:
1. setAlarm is in start(), so cancelAlarm logically belongs in stop(), not delete()
2. It makes more sense to cancel the alarm before deleting the event
3. cancelAlarm being called from delete() meant that if an event doesn't start properly in start(), cancelAlarm got called for no reason.